Update version 2.9.8
git-svn-id: svn://svn.r-forge.r-project.org/svnroot/rbamtools@77 edb9625f-4e0d-4859-8d74-9fd3b1da38cb

+++ pkg/src/samtools/kseq.h
@@ -31,8 +31,6 @@
 #include <ctype.h>
 #include <string.h>
 #include <stdlib.h>
-
-/* Settings for optional embedding into R*/
 #include "rdef.h"
 
 #define KS_SEP_SPACE 0 // isspace(): \t, \n, \v, \f, \r
@@ -134,7 +132,7 @@ typedef struct __kstring_t {
 			str->s = (char*)calloc(1, 1);								\
 		}																\
 		str->s[str->l] = '\0';											\
-		return (int) str->l;											\
+		return str->l;													\
 	} \
 	static R_INLINE int ks_getuntil(kstream_t *ks, int delimiter, kstring_t *str, int *dret) \
 	{ return ks_getuntil2(ks, delimiter, str, dret, 0); }
